# optimize node slices