--- a/match.go
+++ b/match.go
@@ -45,8 +45,36 @@ func (m *matcher) resetCapture(state *MatcherState) {
 	}
 }
 
+func (m *matcher) toStmtSlice(state *MatcherState, nodes ...ast.Node) *NodeSlice {
+	slice := m.allocNodeSlice(state)
+	var stmts []ast.Stmt
+	for _, node := range nodes {
+		switch x := node.(type) {
+		case nil:
+		case ast.Stmt:
+			stmts = append(stmts, x)
+		case ast.Expr:
+			stmts = append(stmts, &ast.ExprStmt{X: x})
+		default:
+			panic(fmt.Sprintf("unexpected node type: %T", x))
+		}
+	}
+	slice.assignStmtSlice(stmts)
+	return slice
+}
+
+func (m *matcher) allocNodeSlice(state *MatcherState) *NodeSlice {
+	if state.nodeSlicesUsed < len(state.nodeSlices) {
+		i := state.nodeSlicesUsed
+		state.nodeSlicesUsed++
+		return &state.nodeSlices[i]
+	}
+	return &NodeSlice{}
+}
+
 func (m *matcher) MatchNode(state *MatcherState, n ast.Node, accept func(MatchData)) {
 	state.pc = 0
+	state.nodeSlicesUsed = 0
 	inst := m.nextInst(state)
 	switch inst.op {
 	case opMultiStmt:
@@ -91,24 +119,32 @@ func (m *matcher) MatchNode(state *MatcherState, n ast.Node, accept func(MatchDa
 }
 
 func (m *matcher) walkDeclSlice(state *MatcherState, decls []ast.Decl, accept func(MatchData)) {
-	m.walkNodeSlice(state, declSlice(decls), accept)
+	slice := m.allocNodeSlice(state)
+	slice.assignDeclSlice(decls)
+	m.walkNodeSlice(state, slice, accept)
 }
 
 func (m *matcher) walkExprSlice(state *MatcherState, exprs []ast.Expr, accept func(MatchData)) {
-	m.walkNodeSlice(state, ExprSlice(exprs), accept)
+	slice := m.allocNodeSlice(state)
+	slice.assignExprSlice(exprs)
+	m.walkNodeSlice(state, slice, accept)
 }
 
 func (m *matcher) walkStmtSlice(state *MatcherState, stmts []ast.Stmt, accept func(MatchData)) {
-	m.walkNodeSlice(state, stmtSlice(stmts), accept)
+	slice := m.allocNodeSlice(state)
+	slice.assignStmtSlice(stmts)
+	m.walkNodeSlice(state, slice, accept)
 }
 
-func (m *matcher) walkNodeSlice(state *MatcherState, nodes NodeSlice, accept func(MatchData)) {
+func (m *matcher) walkNodeSlice(state *MatcherState, nodes *NodeSlice, accept func(MatchData)) {
 	sliceLen := nodes.Len()
 	from := 0
+	tmpSlice := m.allocNodeSlice(state)
 	for {
 		state.pc = 1 // FIXME: this is a kludge
 		m.resetCapture(state)
-		matched, offset := m.matchNodeList(state, nodes.slice(from, sliceLen), true)
+		nodes.SliceInto(tmpSlice, from, sliceLen)
+		matched, offset := m.matchNodeList(state, tmpSlice, true)
 		if matched == nil {
 			break
 		}
@@ -422,11 +458,11 @@ func (m *matcher) matchNodeWithInst(state *MatcherState, inst instruction, n ast
 	case opIfNamedOptStmt:
 		n, ok := n.(*ast.IfStmt)
 		return ok && n.Else == nil && m.matchNode(state, n.Body) &&
-			m.matchNamed(state, m.stringValue(inst), toStmtSlice(n.Cond, n.Init))
+			m.matchNamed(state, m.stringValue(inst), m.toStmtSlice(state, n.Cond, n.Init))
 	case opIfNamedOptElseStmt:
 		n, ok := n.(*ast.IfStmt)
 		return ok && n.Else != nil && m.matchNode(state, n.Body) && m.matchNode(state, n.Else) &&
-			m.matchNamed(state, m.stringValue(inst), toStmtSlice(n.Cond, n.Init))
+			m.matchNamed(state, m.stringValue(inst), m.toStmtSlice(state, n.Cond, n.Init))
 
 	case opCaseClause:
 		n, ok := n.(*ast.CaseClause)
@@ -641,33 +677,43 @@ func (m *matcher) matchArgList(state *MatcherState, exprs []ast.Expr) bool {
 }
 
 func (m *matcher) matchStmtSlice(state *MatcherState, stmts []ast.Stmt) bool {
-	matched, _ := m.matchNodeList(state, stmtSlice(stmts), false)
+	slice := m.allocNodeSlice(state)
+	slice.assignStmtSlice(stmts)
+	matched, _ := m.matchNodeList(state, slice, false)
 	return matched != nil
 }
 
 func (m *matcher) matchExprSlice(state *MatcherState, exprs []ast.Expr) bool {
-	matched, _ := m.matchNodeList(state, ExprSlice(exprs), false)
+	slice := m.allocNodeSlice(state)
+	slice.assignExprSlice(exprs)
+	matched, _ := m.matchNodeList(state, slice, false)
 	return matched != nil
 }
 
 func (m *matcher) matchFieldSlice(state *MatcherState, fields []*ast.Field) bool {
-	matched, _ := m.matchNodeList(state, fieldSlice(fields), false)
+	slice := m.allocNodeSlice(state)
+	slice.assignFieldSlice(fields)
+	matched, _ := m.matchNodeList(state, slice, false)
 	return matched != nil
 }
 
 func (m *matcher) matchIdentSlice(state *MatcherState, idents []*ast.Ident) bool {
-	matched, _ := m.matchNodeList(state, identSlice(idents), false)
+	slice := m.allocNodeSlice(state)
+	slice.assignIdentSlice(idents)
+	matched, _ := m.matchNodeList(state, slice, false)
 	return matched != nil
 }
 
 func (m *matcher) matchSpecSlice(state *MatcherState, specs []ast.Spec) bool {
-	matched, _ := m.matchNodeList(state, specSlice(specs), false)
+	slice := m.allocNodeSlice(state)
+	slice.assignSpecSlice(specs)
+	matched, _ := m.matchNodeList(state, slice, false)
 	return matched != nil
 }
 
 // matchNodeList matches two lists of nodes. It uses a common algorithm to match
 // wildcard patterns with any number of nodes without recursion.
-func (m *matcher) matchNodeList(state *MatcherState, nodes NodeSlice, partial bool) (matched ast.Node, offset int) {
+func (m *matcher) matchNodeList(state *MatcherState, nodes *NodeSlice, partial bool) (matched ast.Node, offset int) {
 	sliceLen := nodes.Len()
 	inst := m.nextInst(state)
 	if inst.op == opEnd {
@@ -727,7 +773,9 @@ func (m *matcher) matchNodeList(state *MatcherState, nodes NodeSlice, partial bo
 		case "", "_":
 			return true
 		}
-		return m.matchNamed(state, wildName, nodes.slice(wildStart, j))
+		slice := m.allocNodeSlice(state)
+		nodes.SliceInto(slice, wildStart, j)
+		return m.matchNamed(state, wildName, slice)
 	}
 	for ; inst.op != opEnd || j < sliceLen; inst = m.nextInst(state) {
 		if inst.op != opEnd {
@@ -776,7 +824,9 @@ func (m *matcher) matchNodeList(state *MatcherState, nodes NodeSlice, partial bo
 	if !wouldMatch() {
 		return nil, -1
 	}
-	return nodes.slice(partialStart, partialEnd), partialEnd + 1
+	slice := m.allocNodeSlice(state)
+	nodes.SliceInto(slice, partialStart, partialEnd)
+	return slice, partialEnd + 1
 }
 
 func (m *matcher) matchRangeClause(state *MatcherState, n ast.Node, accept func(MatchData)) {
@@ -919,58 +969,56 @@ func equalNodes(x, y ast.Node) bool {
 	if x == nil || y == nil {
 		return x == y
 	}
-	switch x := x.(type) {
-	case stmtSlice:
-		y, ok := y.(stmtSlice)
-		if !ok || len(x) != len(y) {
+	if x, ok := x.(*NodeSlice); ok {
+		y, ok := y.(*NodeSlice)
+		if !ok || x.Kind != y.Kind || x.Len() != y.Len() {
 			return false
 		}
-		for i := range x {
-			if !astequal.Stmt(x[i], y[i]) {
-				return false
+		switch x.Kind {
+		case ExprNodeSlice:
+			for i, n1 := range x.exprSlice {
+				n2 := y.exprSlice[i]
+				if !astequal.Expr(n1, n2) {
+					return false
+				}
+			}
+		case StmtNodeSlice:
+			for i, n1 := range x.stmtSlice {
+				n2 := y.stmtSlice[i]
+				if !astequal.Stmt(n1, n2) {
+					return false
+				}
+			}
+		case FieldNodeSlice:
+			for i, n1 := range x.fieldSlice {
+				n2 := y.fieldSlice[i]
+				if !astequal.Node(n1, n2) {
+					return false
+				}
+			}
+		case IdentNodeSlice:
+			for i, n1 := range x.identSlice {
+				n2 := y.identSlice[i]
+				if n1.Name != n2.Name {
+					return false
+				}
+			}
+		case SpecNodeSlice:
+			for i, n1 := range x.specSlice {
+				n2 := y.specSlice[i]
+				if !astequal.Node(n1, n2) {
+					return false
+				}
+			}
+		case DeclNodeSlice:
+			for i, n1 := range x.declSlice {
+				n2 := y.declSlice[i]
+				if !astequal.Decl(n1, n2) {
+					return false
+				}
 			}
 		}
 		return true
-	case ExprSlice:
-		y, ok := y.(ExprSlice)
-		if !ok || len(x) != len(y) {
-			return false
-		}
-		for i := range x {
-			if !astequal.Expr(x[i], y[i]) {
-				return false
-			}
-		}
-		return true
-	case declSlice:
-		y, ok := y.(declSlice)
-		if !ok || len(x) != len(y) {
-			return false
-		}
-		for i := range x {
-			if !astequal.Decl(x[i], y[i]) {
-				return false
-			}
-		}
-		return true
-
-	default:
-		return astequal.Node(x, y)
 	}
-}
-
-func toStmtSlice(nodes ...ast.Node) stmtSlice {
-	var stmts []ast.Stmt
-	for _, node := range nodes {
-		switch x := node.(type) {
-		case nil:
-		case ast.Stmt:
-			stmts = append(stmts, x)
-		case ast.Expr:
-			stmts = append(stmts, &ast.ExprStmt{X: x})
-		default:
-			panic(fmt.Sprintf("unexpected node type: %T", x))
-		}
-	}
-	return stmtSlice(stmts)
+	return astequal.Node(x, y)
 }